Bugfixes for the stepper SM and tune up project setup readme (#7)

+++ src/Motion/MotionController.cpp
@@ -270,7 +270,7 @@ bool MotionController::SetTargetSpeed(AxisLabel anAxisLabel, uint16_t aSpeed)
 	return true;
 }
 
-uint16_t MotionController::GetCurrentSpeed(AxisLabel anAxisLabel)
+float MotionController::GetCurrentSpeed(AxisLabel anAxisLabel)
 {
 	switch (anAxisLabel)
 	{
